refactor(fossid-webapp): Move the `waitDownloadComplete` function
Move this function to the handler as it is specific to the clone repository
workflow.
